Add alternative variational form
The flow law is multiplied by `h` in the minimization form. Try instead
writing out the variational form explicitly but without this factor of
`h` to see if it's numerically a little nicer.

In case we want to still divide through by the thickness formally, we
can instead replace it with a constant value (say 50 or 100m) which will
also get all the units right.