Better conditionals for linear / nonlinear
The old code checked that the float value of `m` or `n` was equal to 1
and dispatched the expression for stress ** exponent. This is evaluated
once when the form is created so we end up with annoying special case
handling on the first step of a continuation method. If we use
`firedrake.conditional` instead then we can get rid of that annoying
first step.