Revert changes to mass balance, use discontinuous velocity
The changes to the mass balance function were making the tests fail
because we don't actually want the factor of `h` inside the `max_value`
that defines the upwind flux. We can get some of the effect using a
discontinuous velocity field but all this does is change the thickness
downstream to its upstream value * (u / (u + u_c)) which is not zero.